Create players and matches tables
Added SQL to create a players table, and a matches table. The players
table has an autoincrementing id, a first name, and a last name. The
matches table has the ids of the two players in the match, and the
id of the winner. The combination of the two players' ids is used as
the primary key in matches.