Implement the playerStandings method
The playerStandings method returns the number of wins and the number of
matches for each player. To do this, I created 3 new views. The playerwins
view shows the number of times each player has won, the playermatches view
shows how many times each player has competed, and the playerstandings view
shows a combination of those two views along with the player's ID and name.